[TE] detection config sermantic validator fix
