+++ thirdeye/thirdeye-pinot/src/main/java/org/apache/pinot/thirdeye/detection/onboard/YamlOnboardingTaskRunner.java
@@ -124,7 +124,7 @@ public class YamlOnboardingTaskRunner implements TaskRunner {
     }
 
     // re-tune the detection pipeline because tuning is depend on replay result. e.g. algorithm-based alert filter
-    DetectionConfigTuner detectionConfigTuner = new DetectionConfigTuner(config, provider);
+    DetectionConfigTuner detectionConfigTuner = new DetectionConfigTuner(config, provider, loader);
     DetectionConfigDTO tunedConfig = detectionConfigTuner.tune(info.getTuningWindowStart(), info.getTuningWindowEnd());
     this.detectionDAO.save(tunedConfig);
